separated useStates/useEffects from AdminCalendar component file
